Misc enhancements to the broker routing (#4831)
- Replace the buffer with wrapped buffer to save the bytes copy
- Change the response size and delay return type to int

+++ pinot-core/src/main/java/org/apache/pinot/core/transport/AsyncQueryResponse.java
@@ -82,8 +82,8 @@ public class AsyncQueryResponse {
     _responseMap.get(serverRoutingInstance).markRequestSubmitted();
   }
 
-  void receiveDataTable(ServerRoutingInstance serverRoutingInstance, DataTable dataTable, long responseSize,
-      long deserializationTimeMs) {
+  void receiveDataTable(ServerRoutingInstance serverRoutingInstance, DataTable dataTable, int responseSize,
+      int deserializationTimeMs) {
     _responseMap.get(serverRoutingInstance).receiveDataTable(dataTable, responseSize, deserializationTimeMs);
     _countDownLatch.countDown();
   }
